requirements.txt
pydantic==2.5.0
streamlit==1.12.0
altair==4.2.2
pandas==2.1.0
numpy==1.24.0
scikit-learn==1.3.0
google-generativeai==0.3.0
jinja2==3.1.2
python-dotenv==1.0.0
plotly==5.17.0
matplotlib==3.7.0
seaborn==0.12.0 